:wrench: [refactor] (#88) Join 부분에 Hilt/코루틴 아키텍처 적용
